pyproject.toml
[project]
name = "cotd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-login>=0.6.3",
    "flask-mail>=0.10.0",
    "python-dotenv>=1.1.1",
    "werkzeug>=3.1.3",
]
